+++ shiny/defaults/settings.R
@@ -1,0 +1,1 @@
+list("username" = "Firstname Lastname", "synchpw" = "password")